# spec/openapi.json
{
  "components": {
    "responses": {
      "APIError": {
        "content": {
          "application/json": {
            "schema": {
              "externalDocs": {
                "url": "https://pkg.go.dev/encore.dev/beta/errs#Error"
              },
              "properties": {
                "code": {
                  "description": "Error code",
                  "example": "not_found",
                  "externalDocs": {
                    "url": "https://pkg.go.dev/encore.dev/beta/errs#ErrCode"
                  },
                  "type": "string"
                },
                "details": {
                  "description": "Error details",
                  "type": "object"
                },
                "message": {
                  "description": "Error message",
                  "type": "string"
                }
              },
              "title": "APIError",
              "type": "object"
            }
          }
        },
        "description": "Error response"
      }
    },
...
  },
  "info": {
    "description": "Generated by encore",
    "title": "API for ewix8",
    "version": "1",
    "x-logo": {
      "altText": "Encore logo",
      "backgroundColor": "#EEEEE1",
      "url": "https://encore.dev/assets/branding/logo/logo-black.png"
    }
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/gitea/public": {
      "get": {
        "operationId": "GET:skyforge.GiteaPublicSSO",
        "responses": {
          "200": {
            "description": "Success response"
          },
          "default": {
            "$ref": "#/components/responses/APIError"
          }
        },
        "summary": "GiteaPublicSSO logs into Gitea with a read-only public account.\n"
      }
    },
    "/api/gitea/sso": {
      "get": {
        "operationId": "GET:skyforge.GiteaSSO",
        "responses": {
          "200": {
            "description": "Success response"
          },
          "default": {
            "$ref": "#/components/responses/APIError"
          }
        },
        "summary": "GiteaSSO performs a one-time token exchange for Gitea and redirects to the UI.\n"
      }
    },
...
    "/api/login": {
      "post": {
        "operationId": "POST:skyforge.Login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "password": {
                    "type": "string"
                  },
                  "username": {
                    "type": "string"
                  }
                },
                "required": [
                  "username",
                  "password"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success response"
          },
          "default": {
            "$ref": "#/components/responses/APIError"
          }
        }
      }
    },
...
    "/api/nautobot/sso": {
      "get": {
        "operationId": "GET:skyforge.NautobotSSO",
        "responses": {
          "200": {
            "description": "Success response"
          },
          "default": {
            "$ref": "#/components/responses/APIError"
          }
        },
        "summary": "NautobotSSO performs a one-time token exchange for Nautobot and redirects to the\nUI.\n"
      }
    },
    "/api/netbox/sso": {
      "get": {
        "operationId": "GET:skyforge.NetboxSSO",
        "responses": {
          "200": {
            "description": "Success response"
          },
          "default": {
            "$ref": "#/components/responses/APIError"
          }
        },
        "summary": "NetboxSSO performs a one-time token exchange for NetBox and redirects to the UI.\n"
      }
    },
...
    "/api/semaphore/sso": {
      "get": {
        "operationId": "GET:skyforge.SemaphoreSSO",
        "responses": {
          "200": {
            "description": "Success response"
          },
          "default": {
            "$ref": "#/components/responses/APIError"
          }
        },
        "summary": "SemaphoreSSO performs a one-time token exchange for Semaphore and redirects to\nthe UI.\n"
      }
    },
    "/api/semaphore/sso/status": {
      "get": {
        "operationId": "GET:skyforge.SemaphoreSSOStatus",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "status": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "status"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Success response"
          },
          "default": {
            "$ref": "#/components/responses/APIError"
          }
        },
        "summary": "SemaphoreSSOStatus verifies that the current user can perform SSO.\n"
      }
    },
...
    "/auth/login": {
      "post": {
        "operationId": "POST:skyforge.AuthLogin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "password": {
                    "type": "string"
                  },
                  "username": {
                    "type": "string"
                  }
                },
                "required": [
                  "username",
                  "password"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success response"
          },
          "default": {
            "$ref": "#/components/responses/APIError"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "description": "Skyforge (deployed behind Traefik)",
      "url": "/api/skyforge"
    },
    {
      "description": "Encore local dev environment",
      "url": "http://localhost:4000"
    }
  ]
}